+++ packages/api-server/src/Server.ts
@@ -4,6 +4,7 @@ import fs from "fs-extra";
 import { BAD_REQUEST } from "http-status-codes";
 import morgan from "morgan";
 import path from "path";
+import { setLogger } from "./core";
 import { baseRouter } from "./routes";
 
 export function appModule({ logPath }: { logPath: string }) {
@@ -16,7 +17,11 @@ export function appModule({ logPath }: { logPath: string }) {
       // @ts-ignore
       return JSON.stringify(req.body);
     });
+    if (fs.existsSync(logPath)) {
+      fs.moveSync(logPath, `${logPath}.old`, { overwrite: true });
+    }
     const accessLogStream = fs.createWriteStream(logPath, { flags: "a" });
+    setLogger({ logPath });
     app.use(
       morgan(
         ":method :url :status :response-time ms - :res[content-length] :body - :req[content-length]",
@@ -25,7 +30,6 @@ export function appModule({ logPath }: { logPath: string }) {
         }
       )
     );
-    //app.use(morgan("combined", { stream: accessLogStream }));
   }
 
   app.get("/api/static", (_req, res) => {