+++ packages/plugin-core/src/logger.ts
@@ -20,8 +20,10 @@ export class Logger {
     const ctx = "Logger:configure";
     fs.ensureDirSync(context.logPath);
     const logPath = path.join(context.logPath, "dendron.log");
+    if (fs.existsSync(logPath)) {
+      fs.moveSync(logPath, `${logPath}.old`, { overwrite: true });
+    }
     fs.ensureFileSync(logPath);
-    fs.truncateSync(logPath);
     let log_level: string;
     if (getStage() === "test") {
       log_level = process.env["LOG_LEVEL"] || "debug";
@@ -102,6 +104,9 @@ export class Logger {
         if (Logger.cmpLevels(lvl, "error")) {
           if (!_.isUndefined(msg?.friendly)) {
             cleanMsg = msg.friendly;
+          }
+          if (!_.isUndefined(msg?.err?.friendly)) {
+            cleanMsg = msg.err.friendly;
           }
           window.showErrorMessage(cleanMsg);
         } else if (Logger.cmpLevels(lvl, "info")) {